48_split examples into provider & client parts
